The naming convention is a bit different here, because we're actually able to go after the conditional total effect directly. That is, we can define a loss function whose minimizer defines the conditional mean of the difference between Qbar_a and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A. We can then define a submodel through this conditional mean difference (which is exactly the conditional total effect) and target this quantity directly.

The naming convention is a bit different here, because we're actually able to go after the conditional total effect directly. That is, we can define a loss function whose minimizer defines the conditional mean of the difference between Qbar_a and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A. We can then define a submodel through this conditional mean difference (which is exactly the conditional total effect) and target this quantity directly.

target_conditional_total_effect(Qbarbar, gn, Qbar, Y, A, a, a_star, ...)

Arguments

Qbarbar
Iterated mean estimates

gn
Power users may wish to pass in their own properly formatted list of the propensity score so that nuisance parameters can be fitted outside of intermed.

Qbar
Outcome regression estimates

Y
A vector of continuous or binary outcomes.

A
A vector of binary treatment assignment (assumed to be equal to 0 or 1).

a
The label for the treatment. The effects estimates returned pertain to estimation of interventional effects of a versus a_star.

a_star
The label for the treatment. The effects estimates returned pertain to estimation of interventional effects of a versus a_star.